requirements.txt
# ── Core Data Science ───────────────────────────────────────
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.0.3
lightgbm==4.3.0
imbalanced-learn==0.12.2
shap==0.44.1
joblib==1.3.2

# ── Visualisation ───────────────────────────────────────────
matplotlib==3.8.4
seaborn==0.13.2
plotly==5.21.0

# ── MLflow ──────────────────────────────────────────────────
mlflow==2.11.3

# ── API ─────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
python-multipart==0.0.9

# ── Database ────────────────────────────────────────────────
sqlalchemy==2.0.29
psycopg2-binary==2.9.9

# ── Dashboard ───────────────────────────────────────────────
streamlit==1.33.0

# ── Notebooks ───────────────────────────────────────────────
jupyter==1.0.0
ipykernel==6.29.4

# ── Testing ─────────────────────────────────────────────────
pytest==8.1.1
httpx==0.27.0

# ── Utils ───────────────────────────────────────────────────
python-dotenv==1.0.1
loguru==0.7.2
tqdm==4.66.2
requests==2.31.0

# ── Fix pyarrow conflict ────────────────────────────────────
pyarrow==14.0.2